recoup/python-backend/requirements.txt
# Recoup Python Backend Requirements

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Firebase & Google Cloud
firebase-admin>=6.5.0
google-cloud-firestore>=2.18.0
google-cloud-storage>=2.18.2

# Database & Caching
redis>=5.2.0
upstash-redis>=1.1.0

# Email
sendgrid>=6.11.0

# SMS & Voice
twilio>=9.3.0

# AI & ML
openai>=1.54.0
deepgram-sdk>=3.7.2

# Payment Processing
stripe>=11.1.0

# Physical Mail
lob>=7.0.0

# Authentication & Security
pyjwt>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12

# HTTP Client
httpx>=0.27.0
aiohttp>=3.10.0
requests>=2.32.0

# PDF Generation
reportlab>=4.2.0
weasyprint>=62.3

# Task Queue & Background Jobs
celery>=5.4.0
apscheduler>=3.10.4
redis>=5.2.0  # For Celery broker

# Utilities
python-dotenv>=1.0.1
python-dateutil>=2.9.0
pytz>=2024.2
nanoid>=2.0.0

# Data Validation
email-validator>=2.2.0

# Logging & Monitoring
sentry-sdk[fastapi]>=2.15.0
python-json-logger>=2.0.7

# Rate Limiting
slowapi>=0.1.9

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
httpx>=0.27.0  # For testing FastAPI
faker>=30.0.0

# Development
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0
isort>=5.13.0
pre-commit>=3.8.0

# Type Stubs
types-requests>=2.32.0
types-pytz>=2024.2.0
types-python-dateutil>=2.9.0

# Documentation
mkdocs>=1.6.0
mkdocs-material>=9.5.0

# Webhooks
svix>=1.37.0

# Environment
python-decouple>=3.8

# Async Support
asyncio>=3.4.3
aiofiles>=24.1.0

# Formatting & Currency
babel>=2.16.0

# Image Processing (for evidence uploads)
pillow>=10.4.0

# Timezone Data
tzdata>=2024.2

# Voice Processing
pydub>=0.25.1
